Long-ify data by bookshelf tags

List each book once for each bookshelf tag it has.

Usage

expand_shelves(data)

Arguments

data

[df] Goodreads data export with "Bookshelves" column.

Value

data [df] Long form of input data with one row per Bookshelf per book.

Details

Tagging books with different bookshelves is a useful way to categorize books on Goodreads. Books can be on multiple bookshelves. This function creates a long version of the data where each bookshelf-book combination is its own row.

Examples

# This example uses our built-in books dataset.
expand_shelves(books)
#> # A tibble: 439 × 25
#>     Book.Id Title    Author Author.l.f Additional.Authors ISBN  ISBN13 My.Rating
#>       <int> <chr>    <chr>  <chr>      <chr>              <chr> <chr>      <int>
#>  1 58885849 Who’s B… Henry… Jr., Henr… "Andrew S. Curran" "=\"… "=\"9…         0
#>  2 11552726 Behind … Susan… Patron, S… ""                 "=\"… "=\"9…         3
#>  3 11552726 Behind … Susan… Patron, S… ""                 "=\"… "=\"9…         3
#>  4 11552726 Behind … Susan… Patron, S… ""                 "=\"… "=\"9…         3
#>  5 11552726 Behind … Susan… Patron, S… ""                 "=\"… "=\"9…         3
#>  6 11552726 Behind … Susan… Patron, S… ""                 "=\"… "=\"9…         3
#>  7   129223 The Wel… Radcl… Hall, Rad… ""                 "=\"… "=\"9…         0
#>  8 51771013 Female … Jen M… Manion, J… ""                 "=\"… "=\"9…         0
#>  9 43980644 Gentlem… Anne … Choma, An… "Sally Wainwright" "=\"… "=\"9…         0
#> 10  9505358 The Sec… Anne … Lister, A… "Helena Whitbread" "=\"… "=\"\…         0
#> # ℹ 429 more rows
#> # ℹ 17 more variables: Average.Rating <dbl>, Publisher <chr>, Binding <chr>,
#> #   Number.of.Pages <int>, Year.Published <int>,
#> #   Original.Publication.Year <int>, Date.Read <chr>, Date.Added <chr>,
#> #   Bookshelves <chr>, Bookshelves.with.positions <chr>, Exclusive.Shelf <chr>,
#> #   My.Review <chr>, Spoiler <chr>, Private.Notes <lgl>, Read.Count <int>,
#> #   Owned.Copies <int>, shelf_list <chr>